requirements.txt
Cerberus==1.3.1
Django==2.2.4
djangorestframework==3.10.2
Pillow==6.1.0
pkg-resources==0.0.0
pytz==2019.2
serpy==0.3.1
six==1.12.0
sqlparse==0.3.0
